Added UI icons and fonts for MKS_UI
Installation: The directories 'mks_pic' and 'mks_font' needs to be put in SD card root (bootloader will update them itself).
Note: There used to be an issue with the bootloader not renaming the directory afterwards (so that printer reboot does reapply the fonts). If this is still present (didn't check), rename/remove the dirs manually after upload.

Image source:
 * https://bit.ly/3hQbsvp (official MKS repo)
 * https://bit.ly/37QH5A4 (startup logo for Sapphire PRO - from official V2.0.0 release package @https://twotrees3dprinter.com)